Rename job from configure-eirini to configure-eirini-scf
At the moment, the configuration step differs in several ways depending
on whether `eirini` is running as a container on SCF or as a job in a
BOSH-deployed environment. To be able to rapidly iterate on both flows
without having incompatible changes, we're splitting the
configure-eirini job into two for the time being. Ideally, we'd be able
to reconcile these jobs back into one in the future.

In a future commit, we will add a configure-eirini-bosh job.

Signed-off-by: Jen Spinney <jspinney@pivotal.io>